pyproject.toml
[project]
name = "search-datasets"
version = "0.1.0"
description = "A collection of search and retrieval datasets for benchmarking"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Zhang Chen" }]
keywords = ["search", "retrieval", "datasets", "benchmark", "RAG", "multimodal"]

dependencies = [
    # Core
    "datasets>=3.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "polars>=1.0.0",
    # File formats
    "pyarrow>=15.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.0.0",
    "pypdf>=4.0.0",
    "python-pptx>=1.0.0",
    # Download & HTTP
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "aiofiles>=24.0.0",
    # Progress & logging
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    # Config & serialization
    "pyyaml>=6.0.0",
    "orjson>=3.9.0",
    "pydantic>=2.0.0",
    # Embeddings & retrieval
    "sentence-transformers>=3.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "faiss-cpu>=1.8.0",
    # Image & multimodal
    "pillow>=10.0.0",
    "opencv-python>=4.9.0",
    # Evaluation
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "pymilvus>=2.6.8",
    "openai>=2.16.0",
    "milvus-lite>=2.4.0",
    "torchvision>=0.25.0",
    "torchaudio>=2.10.0",
    "librosa>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["datasets"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]
